Lesson 6: More Files Update
